Massive Overhaul
Fixed alot of the commands and added better documentation. Also implemented two new HashMaps to keep track of player Teams and player Classes. We use this to prevent two people with the same team to be the same class. Use /teams to see every player and their team, this list resets with server restart!